% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_ta.R
\name{con_ta}
\alias{con_ta}
\title{Contingency Table Generation}
\usage{
con_ta(x, y, level = 5)
}
\arguments{
\item{x}{a numeric vector of data values, each element range from 1 to level.}

\item{y}{a numeric vector of data values, must have same length as x.}

\item{level}{the dimention of the contengency table, the default is 5.}
}
\value{
A contingency table based on \code{x} and \code{y}.
}
\description{
Generate Contengency table for Svensson's Method
}
\examples{
x <- c (1:5,5:1)
y <- c(1:5,1,1,5,4,1)
con_ta(x,y,)
}
\seealso{
\code{\link{sresult}} for summary of Svensson's method analysis.
}
