% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa.R
\name{pa}
\alias{pa}
\title{Percentage Agreement}
\usage{
pa(t)
}
\arguments{
\item{t}{The contingency table for Svensson's method, a two-dimension matrix.}
}
\value{
\code{pa} gives the PA value, multiply by 100 to get a percentage number.
}
\description{
The percentage agreement (PA) which shows the proportion of the subjects who did not change their choices.
}
\examples{
x <- c (1:5,5:1)
y <- c(1:5,1,1,5,4,1)
z <- con_ta(x,y,)
pa(z)
}
\seealso{
\code{\link{con_ta}} for generating contingency table. \code{\link{sresult}} for summary of
Svensson's method analysis.
}
